requirements.txt
boto
click
PyYAML
pyzmq
redis
requests
ujson
